requirements.txt
git+https://github.com/Kucoin/kucoin-python-sdk.git
python-dotenv==1.0.1
requests==2.31.0
python-telegram-bot==13.15
pandas==2.2.2
matplotlib==3.9.0
ta==0.11.0
aiohttp

